remove unstated-next
