+++ package.json
@@ -46,8 +46,7 @@
     "react-router-dom": "^6.2.1",
     "react-use": "^17.3.2",
     "recoil": "^0.6.1",
-    "styled-components": "^5.3.1",
-    "unstated-next": "^1.1.0"
+    "styled-components": "^5.3.1"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0-0",
